+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.0 (unreleased)
 
 - Raise `ArgumentError` for too many arguments with enumerable
+- Removed `uniq` option (use `distinct` instead)
 - Dropped support for Active Record < 5.2 and Ruby < 2.6
 
 ## 0.2.4 (2020-09-07)